Added fix for markers and filters
