+++ js/markers.js
@@ -1,6 +1,11 @@
 /* global L:readonly */
 import {createTemplateElement} from './template-ad-element.js';
 
+const MAINMARKERCOORDIATES = {
+  lat: 35.68950,
+  lng: 139.69171,
+}
+
 const primeMarker = L.icon({
   iconUrl: 'img/pin.svg',
   iconSize: [30, 30],
@@ -13,17 +18,24 @@ const mainMarker = L.icon({
   iconAnhor: [20, 36],
 });
 
-
 const mainPinMarker = L.marker(
   {
-    lat: 35.68950,
-    lng: 139.69171,
+    lat: MAINMARKERCOORDIATES.lat,
+    lng: MAINMARKERCOORDIATES.lng,
   },
   {
     draggable: true,
     icon: mainMarker,
   },
 );
+
+// Функция для возвращения главной метки на исходное место
+const backMarkerToOriginal = (marker) => {
+  marker.setLatLng({
+    lat: MAINMARKERCOORDIATES.lat,
+    lng: MAINMARKERCOORDIATES.lng,
+  });
+};
 
 //Функция для передачи адреса метки в поле ввода
 const getMarkerAdres = (marker, element, decimalPlaces) => {
@@ -59,4 +71,4 @@ const pinGroupToMap = (group, map) => {
 const removeMarker = (map, group) => {
   group.remove(map);
 }
-export {mainPinMarker, createGroupMarks, getMarkerAdres, removeMarker, pinGroupToMap};
+export {mainPinMarker, createGroupMarks, getMarkerAdres, removeMarker, pinGroupToMap, backMarkerToOriginal};
